Cycle terminals in "visible" order
Previously the different terminals of a session have been visited in the order
defined by the iterator of m_terminals. This was rather counter-intuitive as
terminals have been added to the session in the order of their creation. Assume
the session has

(1) first been splitted vertically and then
(2) the left terminal again has been splitted horizontally

resulting in a layout like this:

|--------------------------------------------------------------------------|
|                                      |                                   |
|  terminal: 1st                       | terminal: 2nd                     |
|  visited:  1st                       | visited:  2nd (now 3rd)           |
|                                      |                                   |
|---------------------------------------                                   |
|                                      |                                   |
|  terminal: 3rd                       |                                   |
|  visited:  3rd (now: 2nd)            |                                   |
|                                      |                                   |
|--------------------------------------------------------------------------|

The terminals have been created in the order indicated by "terminal:".
"visited:" displays the order these terminals are visited when switching among
them using the appropriate shortcut, i.e. after visiting terminal "1st" the
cursor jumps to right to terminal "2nd" and finally the cursor jumps back to the
left to terminal "3rd".

This is changed so that terminal "3rd" is visited before terminal "2nd",
i.e. the cursor will jump back from the right side to the left side any more.